vector_rag/pyproject.toml
[tool.poetry]
name = "vector_rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
langchain-community = "^0.3.18"
faiss-cpu = "^1.10.0"
langchain-ollama = "^0.2.3"
langchain-huggingface = "^0.1.2"
langsmith = "^0.3.15"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
